feat(swap): merge cancel/refund commands into one command
