+++ .github/workflows/ci.yml
@@ -134,6 +134,7 @@ jobs:
             happy_path_restart_bob_before_xmr_locked,
             happy_path_restart_alice_after_xmr_locked,
             alice_and_bob_refund_using_cancel_and_refund_command,
+            alice_and_bob_refund_using_cancel_then_refund_command,
             alice_and_bob_refund_using_cancel_and_refund_command_timelock_not_expired,
             punish,
             alice_punishes_after_restart_bob_dead,
