--- swap/src/cli/refund.rs
@@ -1,43 +1,0 @@
-use crate::bitcoin::Wallet;
-use crate::protocol::bob::BobState;
-use crate::protocol::Database;
-use anyhow::{bail, Result};
-use std::convert::TryInto;
-use std::sync::Arc;
-use uuid::Uuid;
-
-pub async fn refund(
-    swap_id: Uuid,
-    bitcoin_wallet: Arc<Wallet>,
-    db: Arc<dyn Database>,
-) -> Result<BobState> {
-    let state = db.get_state(swap_id).await?.try_into()?;
-
-    let state6 = match state {
-        BobState::BtcLocked { state3, .. } => state3.cancel(),
-        BobState::XmrLockProofReceived { state, .. } => state.cancel(),
-        BobState::XmrLocked(state4) => state4.cancel(),
-        BobState::EncSigSent(state4) => state4.cancel(),
-        BobState::CancelTimelockExpired(state6) => state6,
-        BobState::BtcCancelled(state6) => state6,
-        BobState::Started { .. }
-        | BobState::SwapSetupCompleted(_)
-        | BobState::BtcRedeemed(_)
-        | BobState::BtcRefunded(_)
-        | BobState::XmrRedeemed { .. }
-        | BobState::BtcPunished { .. }
-        | BobState::SafelyAborted => bail!(
-            "Cannot refund swap {} because it is in state {} which is not refundable.",
-            swap_id,
-            state
-        ),
-    };
-
-    state6.publish_refund_btc(bitcoin_wallet.as_ref()).await?;
-
-    let state = BobState::BtcRefunded(state6);
-    db.insert_latest_state(swap_id, state.clone().into())
-        .await?;
-
-    Ok(state)
-}
